requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chumpy==0.70
contourpy==1.0.6
cycler==0.11.0
flatten-dict==0.4.2
fonttools==4.38.0
freetype-py==2.3.0
frozenlist==1.3.3
fsspec==2022.11.0
idna==3.4
imageio==2.22.4
joblib==1.2.0
jpeg4py==0.1.4
kiwisolver==1.4.4
loguru==0.6.0
matplotlib==3.6.2
multidict==6.0.3
networkx==2.8.8
numpy==1.23.5
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==22.0
Pillow==9.3.0
protobuf==3.20.1
pycparser==2.21
pyglet==2.0.1
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
smplx==0.1.28
threadpoolctl==3.1.0
tifffile==2022.10.10
tqdm==4.64.1
trimesh==3.17.1
typing_extensions==4.4.0
urllib3==1.26.13
yacs==0.1.8
yarl==1.8.2
numba
filterpy
tensorboard
kornia
pytorch-lightning==1.5.9
torchvision
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git
